PMA-127 - Visitor related policy

+++ app/Http/Controllers/VisitorArchiveController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\VisitorArchive\UpdateRequest;
 use App\Http\Resources\VisitorArchiveResource;
 use App\Http\Resources\VisitorArchiveResourceCollection;
 use App\Repositories\Contracts\VisitorArchiveRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 
 class VisitorArchiveController extends Controller
@@ -32,9 +33,12 @@ class VisitorArchiveController extends Controller
      *
      * @param IndexRequest $request
      * @return VisitorArchiveResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [VisitorArchive::class, $request->get('propertyId')]);
+
         $visitorArchives = $this->visitorArchiveRepository->findBy($request->all());
 
         return new VisitorArchiveResourceCollection($visitorArchives);
@@ -45,9 +49,12 @@ class VisitorArchiveController extends Controller
      *
      * @param  StoreRequest  $request
      * @return VisitorArchiveResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [VisitorArchive::class, $request->get('propertyId')]);
+
         $visitorArchive = $this->visitorArchiveRepository->save($request->all());
 
         return new VisitorArchiveResource($visitorArchive);
@@ -58,9 +65,12 @@ class VisitorArchiveController extends Controller
      *
      * @param VisitorArchive $visitorArchive
      * @return VisitorArchiveResource
+     * @throws AuthorizationException
      */
     public function show(VisitorArchive $visitorArchive)
     {
+        $this->authorize('show', $visitorArchive);
+
         return new VisitorArchiveResource($visitorArchive);
     }
 
@@ -70,9 +80,12 @@ class VisitorArchiveController extends Controller
      * @param UpdateRequest $request
      * @param VisitorArchive $visitorArchive
      * @return VisitorArchiveResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, VisitorArchive $visitorArchive)
     {
+        $this->authorize('update', $visitorArchive);
+
         $visitorArchive = $this->visitorArchiveRepository->update($visitorArchive,$request->all());
 
         return new VisitorArchiveResource($visitorArchive);
@@ -83,9 +96,11 @@ class VisitorArchiveController extends Controller
      *
      * @param VisitorArchive $visitorArchive
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(VisitorArchive $visitorArchive)
     {
+        $this->authorize('destroy', $visitorArchive);
         $this->visitorArchiveRepository->delete($visitorArchive);
 
         return response()->json(null,204);
